Altered QuestionSelectionItem to stop overlap
There was a problem with the QuestionSelectionItem class that only revealed itself on the Vocabulary screen, and only on my phone (and I guess other phones with certain physical dimensions). If the contents string was of a certain length, the text would end up in the same space as the checkbox, overlapping it. I tried to fix this by telling the contents label how big the checkbox was, through the measureChildWithMargins method, which had a parameter called widthUsed. This worked... except... In the circumstance where the height of the contents label before the change was different than the height of the contents label after the change, the text on the last line, moved there by the word wrap, would actually disappear, as well as the tapping effect background on that last line. The reason this was happening? My best guess is that there were multiple internal methods measuring the size of objects, and as a result my change only affected the calculation of the external dimensions, not the dimensions of the child objects. So I was at a loss, until I realized I could just do something I was more comfortable with anyways. I could change the class from extending the RelativeLayout class, and instead extend the (in my mind more predictable) LinearLayout class. So I did. I considered having nested layouts to handle the TextViews, but I realized I could just use HTML in a single TextView to handle the formatting and layout of the text instead. I also realized I did not need to override the onMeasure method anymore, so I scrapped that. The end result is a system that's simpler, prettier, more workable, and overall just better. Very happy.