Added non-breaking spaces to QuestionSelectionItems
The QuestionSelectionItems for vocabulary categories was word wrapping in a less-than ideal way. Because sometimes the break is in the middle of a two-word English translation. Normally this would be fine, but it's splitting up what is essentially a single item, and that affects readability. As one's looking at the list of items, skimming through, they may be looking for a specific question, such as 'Post Office'. But if all they see is 'Office' on one line, and 'Post' on another, they may miss it. So for the sake of readability, I made a minor change. Now, spaces in the middle of items have been replaced with non-breaking spaces to preserve readability. I think this is much better.