Set limiter in multiple choice button size
Was annoyed by the normalizing button size function I designed (abaec48d11f5b8f5b861f73f7604d2b43e5bbf2b) causing the buttons to be a bit too big. So big that it results in there being a mere one button per line, which isn't exactly aesthetically pleasing, to say the least. So, I decided to set a limiter, preventing the buttons from growing beyond half the width of the pad, which means there won't be fewer than two buttons per line. The one problem with this is that in cases where the limiter is necessary, the button that is bigger than half the screen would end up with its text wrapping across two lines, so its height is different, which causes a new inconsistency. Not sure how to deal with that yet.