Altered multiple choice buttons for height consistency
Because of the the size restriction code I added recently (863ff7113304860fe824d73425317513bf0f0126), which was causing the text of some buttons to word wrap when their width was restricted, those buttons were slightly taller than the others. Attempts to fix this after the wrapping was triggered didn't exactly work, and would just force the problematic buttons to be the same, smaller height as all the others, forcing part of the last line of text to disappear by falling outside edge of the button. So I had a thought: Why not just force the multi-word answers to word wrap from the start? That way, I can measure the bigger buttons and set all the smaller buttons to be the same, without worrying about the measured size changing in the middle of the operation. So I did that, replacing the spaces in the button text with line breaks. After that, I could just duplicate the code that I was using for width, rename a few things, rename the maxSize variable so it now refers specifically to width, and I'm done. I debated whether I still needed the limiter code, preventing the buttons from being too wide. I decided to remove it since it would likely cause more problems than it would solve now. Because if it was triggered by a single massive word, the word would actually be unceremoniously broken by a line break in the middle of it, and the last line of text would actually be cut off by the edge of the button, as mentioned earlier. In that circumstance, one button per line is actually preferable, aesthetically.