Modified logGraph style
Made some aesthetic changes to the log graph. The zero on the Y-axis was being cut off by the bottom edge of the GraphView object, so I knew I had to fix that. And after a ton of messing about and trying different things that didn't work, including adjusting the padding which had absolutely no effect, I found that setting a title of one space on the X-axis had the desired effect. I also tweaked the width of the label space on the Y-axis, so it hits the absolute lowest value it could without cutting off the any of the labels.