Fixed the spacing below the X-axis
The spacing below the X-axis on the log graph was a bit too big. But before the most recent change (dd860fc56289a563ee1a6d7958137871e5600a53) where I added a title to the X-axis that was just whitespace, it was too small and was cutting off the zero label of the Y-axis. So I needed to find a solution that was somewhere in the middle. I tried using the setLabelHorizontalHeight method, but it was completely unresponsive. Took me a while to realize, it would only work if there were axis labels, and an axis title wasn't good enough. So I set a couple X-axis labels that were pure whitespace, which it turns out I can do using the StaticLabelsFormatter, and used the setLabelHorizontalHeight method to set the spacing to the lowest value I could without cutting off the zero. Looks a lot better now.