Updated testing library
The Espresso library, used for instrument tests, has finally come out of beta on AndroidX, so I updated it to the latest version. I also did a bit of correction, because there were deprecated items that needed to be replaced. But one of the items, the replacement for the JUnit4 runner, wasn't included in the newest packages, not even as a dependency, so I had to add the relevant library to the build file. It's possible I am expected to use a completely different runner, but for now, this'll do.