Added countries to vocabulary
Added several countries to the vocabulary file. Been thinking of adding a few more, using the stats from Firebase to establish where my audience is, and adding their home nations. But there are a lot. This will do for now. These particular countries were listed in Learn Japanese to Survive: Hiragana Battle, so it's a decent start.