Adjusted KanaSet storage for efficiency
The storage of kana sets was way too complicated. It was a four-dimensional array, which in many circumstances was mostly blank. For vocabulary questions it was allocating eight times as many addresses than necessary, because there was no use of diacritics or digraphs. So I decided to reorganize it into a simple Map collection that just stored the sets indexed by a specially designed class. I figure this is much more efficient.