Updated translator request
The request for volunteer translators has been altered. I specified that language codes have to be from the ISO 639-1 specification, and added a link to Wikipedia listing all of them. I also changed the vocabulary request to use the vocabulary.xml file from the 'develop' branch, because it will, in all cases, be more recent; and for consistency I changed the slashes and formatting of the directory name where the translated vocabulary.xml file should be placed.