Rewrote database migration code
Found an SQL command, which I didn't know existed, to quickly and simply migrate the contents of one table to another. So while before, I was using a loop to transfer the contents of one table to another, one by one; this new code uses one line to tell the database to do it all on its own. This saves a lot of space in the code, and a lot of effort in the Java processor. And when it comes to this task, I assume the developers of SQLite know what they're doing far better than I do.