Enhanced encapsulation of database record classes
The various record classes had public parameters, which was just not on. So I modified the classes to use encapsulation, with private parameters and specialized 'get' methods. I also wrote special mutator methods that are specific to the tasks at hand, and just add to or increment the modifiable properties. I also set the primary keys to be final, because those should never be modified once set, and because I didn't write any setters, I instead wrote full parameterized constructors for the Room framework to utilize. End result is a much more robust set of database classes.