Fixed problem with the database
The database was being populated with the wrong primary keys. Instead of using the romaji form of the vocabulary questions, as I intended, it was being populated using whatever form the contents were displayed in, which by default was kana. This is because while I remembered to set up the QuestionBank class to pull data from the database using the correct primary keys, I forgot that the database is populated by the MainQuiz class. I fixed it by having all calls to populate the database call a new method in the QuestionBank class that actually retrieves the database keys properly. I'm putting this change through in a hotfix, because I know it needs to be fixed now, before too much incorrect data is inserted by the user, and I don't have time to fully troubleshoot and cleanup the changes made in the development branch.