Started log details feature
Started work on a new feature, which is the ability to tap on a DailyLogItem, and instantly look the individual scores of individual questions of that day. This is still very much a work in progress. Mainly because I really don't like the appearance of anything at all at this point. But it functions, which is a start. Most of the code in the LogDetailView class is copied verbatim from the LogView class, but altered to fit this new function.