Restyled Bar Chart
The bar chart in the LogDetailView class didn't look right at all. So after experimenting for quite a while, I finally figured out what should go where to make the chart look exactly how I want it. To start, by adding 0.5 to the X value of each data point, I managed to center the bars in their appropriate positions. Next, there was the setSpacing and setDataWidth methods, which I used to appropriately size the bars. But it didn't work at first. Turns out, because I was adding the graphSeries object over and over again in a loop, it was counting all of them, and shrinking the bars to help them fit, which was not what I wanted. I noticed this when I inserted a timing delay in the process loop, and watched what happened each time the series was added. So basically, to solve this, I had to clear everything out of the chart before readding the updated graphSeries object. And now, the bars at least are finally rendered correctly.