Fixed label renderer
I accidentally had the label renderer for the detail chart implemented twice in two different ways. Once using the in-built StaticLabelsFormatter, which was in turn overriding the other one, my custom label formatter, built using an anonymous class. This explains why the X-axis labels kept shifting around a lot in ways I could not understand. So I removed the use of the StaticLabelsFormatter, allowing my custom formatter to take over. But in the process of testing this, I learned my custom formatter was incorrectly built. I got the axes mixed up. So I also fixed that little glitch.