Added padding to chart scrolling
Added two items, intended to be invisible, to the graphSeries object, to allow padding. Because zooming and scrolling is enabled on the horizontal, If the user zooms in, and zooms out on the chart, once they zoom back out, the limits built into the scrolling function will cut off half of both the first and last bars. I found the solution was to add items at the beginning and end of the list, to tell the class that those two positions should still be viewable.