Separated graph padding
The padding items I just created (ffca19bac91f4dd6579db3ae553f1cd90dd8bdf0) to deal with scrolling problems were integrated into the graphSeries object, which didn't feel fit to purpose. That's supposed to store the referenced items, and nothing else. So I created a new series object to handle this function, and added it to the chart separately. But instead of using a BarGraphSeries object, which would cause the bars to shrink in size, as I found out earlier (19b40de7949e0d2928c5a85fc5121f5e910d8273), I used a PointsGraphSeries, and set the size of the points to zero. This feels much better, in terms of functionality, with all the necessary code in the same section. I like it.