Corrected axis label positioning
The positioning of the labels on the X-axis was quite nuts. It would repeat continuously across every point below the given bar. Because the labels appear at every line on the horizontal grid, I corrected this by setting the grid lines to be double the number of graph bars, so grid lines appear in the middle of the bars. And then I set the labels to only appear in the middle of the bars. This does screw up if there's any zooming, but one step at a time.