Fixed issue with improperly-sized TextViews
The TextView objects within the items in LogDetailView were not being sized correctly, and in many cases, the text was being cut off. I quickly realized what was wrong. Turns out, I couldn't just reuse the LayoutParams objects used to position the text within. I'm guessing the size calculation was being done within the LayoutParams objects, and because I was reusing them, a change in one was influencing the size of all the inadvertently connected TextView objects. So by moving the objects to exist within the processing loop, it'll instead create new LayoutParams objects every time. Not quite as efficient, but preferable to the alternative.