Created special LogDetailItem class
Created a new LogDetailItem class for the LogDetailView, replacing the RelativeLayout objects filled with TextViews. This new class is based entirely on DailyLogItem, but modified to fit this new purpose, with the dates replaced with questions, the onClickListener and XML attributes removed, and the code simplified slightly. I might have to go through and find a way to deal with the repeating code, but for now, this will do. I'm also debating whether I need the percentage view in the LogDetailView, but I figure it's easier to remove it later, than add it in later. So it'll stay for now.