Added question file attributes to styleguide
The code reformatter in Android Studio was rearranging the attributes in the question XML files to fit an alphabetical order, like it does to all other files. But I prefer it in a more intuitive order of having the answer to a question at the very end, and having the most important attributes at the very beginning. So I found a way to override this function by adding some of the attributes to the arrangement section of the XML styleguide. That way I don't have to override it manually every time I run the code reformatter.