Switched out graph library
Sick and tired of all the problems I was experiencing with the bar graphs from the GraphView library, I decided to change it out with the MPAndroidChart library. It handles bar charts in a much more intuitive way, and requires less faffing about, specifically with regards to labels on the X-axis. I think this is mainly because the GraphView library tries to handle all types of graphs in the same way, using the same GraphView class for all implementations, but the MPAndroidChart library handles things differently, using specific view classes for specific types of charts. So there's no attempt to combine the functionality of line graphs and bar graphs in the same view. There's also no weird graphical glitches. Overall, it's just a better library.