Added background to DailyLogItem
Because the DailyLogItem objects are now selectable, I decided they should respond visually to taps, like the QuestionSelectionItems (ce60d6b6044a161eb5eb477bfad1b7648f950629). So I added the same background drawable to the class. I ran into one major problem, and that is the getDrawable method is expecting a resource id to be passed. Except it doesn't work that way in a TypedArray with only two items in it, which is the case here since I'm using the same defaultAttributes object I use to fetch the default text colour. I don't know why it presumes otherwise. But the only way I could get it to compile without throwing lint errors is by explicitly suppressing that one error. I tried a workaround involving storing the drawable and reusing it, but then the effect only appears on the last item, even when I'm tapping other items. So I have to create a new drawable object for each one, and this seems to be the only way to do it.