Changed ids to be unique
Not that it's a major problem, but I just thought that the ids defining the objects in LogDetailView should be unique, but they weren't, since I just copy-pasted the layout file from LogView. So I went through the effort of changing all the object ids to be unique to the LogDetailView class. I also changed all uses of the word 'graph' to 'chart.' Since that's what it is now. And yes, there is a difference.