Updated changes to the branching model
Originally, I was planning on posting all changes to the readme and privacy policy to the master branch, because by default that's the branch that appears on GitHub. But I found out I could change the default branch to the develop branch, which is where most of the work is being done. So because of that little revelation, I decided to make that change, and in future post all readme and privacy policy changes to the develop branch instead. Thankfully I haven't made any changes to the readme since the last hotfix (e44c43b3850f6874c4bb29344e8bd078d79aa247) so there is no need to transfer anything new from the master branch. But as part of that, I thought I should update the section of the readme that explains the branching model.