Changed correctAnswers to int
The correctAnswers property in LogDetailItem is now stored as an integer. Because unlike in DailyLogItem, where most of the code has been copy-pasted from, the correctAnswers number is always going to be an integer, because that's how they're stored in the database. So storing it as a float is just unnecessary. As part of this, I also removed the parseCount method that runs the number through the Fraction class, because it's only needed for floats.