Added Kanji to upcoming features
I realized that the vocabulary questions were not the same as proper kanji support, and that they're not interchangeable. The individual kanji can be combined with other kanji and kana to form various words, and kanji are not always pronounced the same way. So proper kanji support will be necessary. I'm going to need a special KanjiQuestion class to support this. But for now, I just added a note to the readme.