Began work kanji questions
Started adding kanji support to the app. The very first step was adding questions. So I took the first five lessons from Learn Japanese To Survive: Kanji Combat, and placed them in an XML document. However even this step is not completed, because the complexity of the kanji means that there are multiple elements to keep track of, and multiple answers for each one. But all the necessary data is in the system, now I just need design a way to file it correctly. Still, it's a good first step.