Kun'yomi and on'yomi are no longer required
Learned that not all kanji have kun'yomi pronunciations, which threw me for a loop. So I decided not make any of the pronunciations a requirement for the kanji questions. The only requirement is that each has a meaning, which I would hope all kanji have. If not, I'm giving up on this whole project.