Redesigned how kun'yomi and on'yomi are stored
The kun'yomi and on'yomi are now stored not as attributes, but as elements within the KanjiQuestion objects. This allows a multiple number of kun'yomi and on'yomi to be stored individually. I could've stored it in a similar way I stored the AltAnswer elements in the KanaQuestion and WordQuestion elements, and still kept at least one of each as attributes to be recognised as the primary kun'yomi and on'yomi, but I thought it wasn't the best way to handle it, as I don't see one kun'yomi or on'yomi offering actual priority over the alternates. So storing them as elements is just a better way to handle them.