Requiring kun'yomi and on'yomi use
Made a tiny change to the XSD document to now specify that at least one of the KunYomi or OnYomi elements must now be included in each KanjiQuestion element. This doesn't affect how many KunYomi or OnYomi elements are actually in the KanjiQuestion, or stop it from having only KunYomi elements or only OnYomi elements. But it does mean that there has to be at least one of one of them. Basically, if the KanjiQuestion element is empty, only then are errors thrown.