Added several dozen more kanji
Added an additional nine kanji sets to the app, all sourced from Learn Japanese to Survive: Kanji Combat. I haven't come up with appropriate names for any of them. Mainly because some of the sets are so disparate that finding a single word or sentence to describe them is not a simple task, and also I cannot be asked right now to do any of that. It can wait.