Got the kanji into the quiz
Finally got the kanji actually added to the app, rather than just sitting in an unused XML file. All I did was add them to the selection screen, and got the QuestionBank class to properly handle the multiple choice features. Right now, the multiple choice feature treats them the same as WordQuestions, which is probably fine, but I might want to change it out so they're handled separately. I'm not sure yet. I'm also thinking about how I'm going to handle the kun'yomi and on'yomi, because I'm quite conflicted on the best way to handle them, either as separate questions with a separate class, or as an alternate function of the same object. It's staying on the back burner for now.