Redesigned kanji answer check
The code that checks if a kanji answer is correct has been enhanced. Because I want the app to detect a partial answer as a correct one, mainly because the original keyboard entry method is still a usable thing, I had it split up the multi-part meanings and run through them to check if the user's response matches one of those sub-answers. If there's a match, the answer's correct! Otherwise, no.