Added kanji to reference screen
Added the kanji to the reference screen. As part of this, I enhanced the ReferenceCell class, so that if a description contains slashes, it'll treat those slashes as separators, and put each individual element on a new line. Makes all the items fit a bit better. There are still a few problems with getting all the items to fit on the width of the screen, and I'm also having a problem with the elements in each set not lining up whenever the count changes. But this is a good start.