Redesigned layout of kanji reference
To deal with the issue of items not lining up properly when the number of items in a given row changes, I rewrote the getKanjiReferenceTable method to, whenever the count changes, start a new table and populate that instead. This makes everything nice and centered, which is much more visually appealing.