Enabled wrapping on spaces in ReferenceCell description
Expanded the code for determining line breaks in the ReferenceCell description by having it also split on spaces if the line gets too long. I specifically set it to break if the line is more than 10 per cent wider than the subjectWidth, to prevent the hook from being too sensitive. As part of this, I also altered the initial line break code to preserve the slashes that were originally the only line break points, in order to differentiate those points from the points where the line breaks occur at spaces. All this is mainly because one item, with the meaning of 'village section,' was so wide it really messed up the layout of the entire kanji reference screen. This fixes it.