Added line breaks to buttons
I was having a problem with buttons in the MultipleChoicePad getting too wide with some of the KanjiQuestions. So, just like the last time I had this problem (6697b6d32e8e7f866ccf629d1c26547fb6b36899), I decided to fix it with a few line breaks, this time placing them wherever there's a slash.