Set slash as a constant
The slash is used to separate multiple meanings in the kanji.xml document. I'm starting to think that works well, and I should stick with it. But I decided to, rather than have it set as literals sprinkled throughout the app, to set it as a constant, like I should've done from the very beginning.