Added some more feature ideas
Added a note to the readme that I plan to add support for kun'yomi and on'yomi at some point. But right now, as I'm in the middle of implementing more basic kanji questions, I'll consider that a separate feature for implementation at a later date. I also had the idea to implement some type of limitation. Because right now, the user can keep going, answering questions until the end of time, without the app making any attempt to stop them and say "I think you've had enough," or at the very least, suggesting they take a break. Some way of getting the user to slow down, do something else. Which may seem like an appalling thing for an app developer to suggest, but I care about my users, and their mental well being. It probably won't tell them to stop, but it'll give some type of subtle nudge that this is a good time for the session to end. And the idea is to also congratulate them for spending that much time, and we could congratulate them with a score of some description. I'm not sure yet.