Redesigned digraphs declarations
The digraph sections in the hiragana and katakana files were being defined using booleans, under the assumption that all of them use kana from the Y-group as the second character. But I've been looking into other types of digraphs, and found this might not always be the case. So I decided to redefine how digraph sections are identified and handled, by having the definitions specify which kana set they're paired with. Currently, this pairs them all with set nine, and none of the code has been modified to examine any deviation in that, but this is merely the first step. I plan to rewrite a lot of the code in QuestionManagement to take full advantage of this change.