Created second kanji file
Added a few more kanji, but because the first file was getting big enough, I decided to make a second file called kanji_2. As part of this, I also segregated the contents of the first file from the second on the Reference and QuestionSelection screens, using tabs and descriptive section headers respectively. Also renamed the first file to kanji_1 for the sake of consistency.