Corrected tab issue
The addition of the kanji category caused the tabs labelled 'hiragana,' 'katakana' and 'vocabulary' on the QuestionSelection and Reference screens to wrap onto two lines with improper line breaks. I fixed it by setting the TabLayouts to be scrollable, much like I did when I encountered the same problem with the vocabulary reference screen (c2eb24fb414b87ede6b4ecb66fa9620eb5132a10). And much like that issue, it's only partially effective, and based on me guessing what the screen size is. What I need to do is find a way to detect when there's an issue like this, and only then make the TabLayout scrollable. But I'm putting that on the back burner. This quick fix will work for now.