Fixed XmlParserTest
The recent redesign of the Section tags (315d33a50039c53ef2257aa4b7bb64a05229321d) broke XmlParserTest. This is because I forgot to update the test to reflect the change, which I needed to because it was a fundamental change in how the diacritic definitions are stored. So I fixed that.