Added an additional set of katakana
There's a special set of katakana that can be used for loanwords containing syllables that are not usually in Japanese. I found a few sources of this. First from Learn Japanese to Survive: Katakana War, second from a Japanese textbook I have, and third, from a random website I ran across. And after removing several duplicates, this is what I'm left with. I might need to verify some of these, reorganize and trim the fat, but for now, as the most complete list, it'll do. I'll trim it down later.