Removed standalone diacritics
The standalone diacritic characters have been removed from the kana datatype in the XSD document. This is mainly because for this kind of thing, one should only be using the combining diacritic marks instead, if one has to use them at all. The standalone ones really have no purpose here.