Wrote RomajiConversion code
Wrote a bit of code to convert any string of kana to the romaji equivalent. Currently uses the 'default' system of romanization, which is basically just Hepburn, but with Nihon-shiki for the W-Set. Got a few ideas of what I'm going to use this for, but for now, just writing the code is enough.