Added macrons to RomajiConversion
Double vowels should be signified with macrons instead. So with use of the replace method I was able to replace every double vowel, as well as 'ou' and 'ei' and use of the chōonpu (or 'Prolonged Sound Mark'), with the appropriate long vowel. Not sure of chaining the method calls was the best way to go about this, but it works.