Fixed the 'ch' doubling
When a sokuon is followed by the 'chi' character (ち/チ) the consonant doubling supposed to be done slightly differently. Because 'chi' is actually part of the T group, we're just supposed to place a 't' before the 'ch', not go 'cc' or 'chch'. So I added a simple hook to deal with that specifically.