Moved LogDetailView title to resources
The title of the LogDetailView was placed as a literal in the class file. This is not how it's supposed to be done. So I moved it to the strings resource file. As part of this, I decided to specify the formatting of the date in the string resource, so any translators can modify it (if I ever get any translation). But I ran into a major problem. The formatting requires using Java's built-in date classes, and I've been avoiding the ones in java.util because they're pants. But this could include the java.time packages, which I cannot use because they don't work on anything older than Oreo (APK 26). Which is why I've been using the ThreeTen backport. But you'd think, since they are basically the same classes, the two would be compatible. Not so much. Because it's specifically looking for java.time, org.threeten.bp is not recognised as the same thing. However the formatting framework will accept a long in lieu of a Date object. So after a bit of research, I got the LocalDate object to convert to a long, counting the number of milliseconds since the Unix epoch. This works. But if and when I decide to ditch any users running Nougat (APK 25) or earlier, I'll have to remember to correct this little nugget.