Fixed font issue
There's a bit of a problem with fonts and kanji. The default fonts were actually rendering the Chinese (Hanzi) versions of some characters, which are markedly different to the proper Japanese versions. I tried several solutions to fix this, the big one being manually importing the Japanese-specific font (Noto Sans JP) and using that. I tried doing this by utilizing the built-in font downloading system, but for some reason that particular font, despite being listed on the Google Fonts website, wasn't available through the Google Fonts API used by Android. So that was a bust. The other method I tried involved embedding the fonts in the app itself. This worked, but ended up ballooning the app to over 20 MB, which is just unacceptable. Then, after discussing the issue on some web forums (the same forums that alerted me to this issue even existing) I had an epiphany: Seems one can just change their locale settings to display typefaces specific to their locale (https://steamcommunity.com/app/759440/discussions/0/1735462352469312874/#c1751232561632906682). And it made me realize something else: Why I was only experiencing this issue on the emulators, and not my actual phone. I thought it was because Samsung was just better, but no, it was because I set Japanese as the number two language on that device. I didn't even touch the locale settings on the emulators and that means by default they would render the characters in Chinese. So it seems the solution to this problem was simpler than I thought. So I looked through the JavaDocs for the TextView class, noticed the setTextLocale method, and realized that by setting the locale to Japanese on any and all TextViews and TextPaints that are used to display kanji, all my problems would be solved. So that's exactly what I did. Get in!