Corrected font locale
As part of my work on adding kanji to the app, I learned about a problem with certain kanji being rendered incorrectly as their chinese equivalents. So in conjunction with my fix on that branch (b53b00dd5a3788c47c5d1bd00f6638ce9ad62c44), I decided to apply an identical fix to the LogDetailItem class in this branch. It's a minor fix that does nothing now, but will save trouble when everything gets merged together.