Altered macron handling
Altered the RomajiConversion code so it no longer places macrons on vowels in katakana when there are double vowels, because I learned that in romanization of katakana you only add macrons for long vowels, and those only occur if there's a Chōonpu following the vowel, not if there are double vowels. Double vowels are different. So I removed all the replace method calls for double vowels at the end, and just had the main bulk of code look for cases of hiragana double vowels that need to be replaced, and mark them with Chōonpus so the replace chain at the end can handle the rest.