Added some more extended katakana
Found additional extended katakana on Wikipedia. Really need to sort through and trim it down because there is a lot here. Also, found a few conflicts, so I now have to determine which one is correct. Is it Wikipedia, or Learn Japanese to Survive? Don't know yet. Pretty sure the Living Language book is right though.