Moved dimensions to resources
There are several points in the Java code where items are resized to dimensions are explicitly specified in Java. This has worked quite well, but really isn't best practice. So I moved almost all of the explicit dimensions to the dimension resource file. The only ones I didn't move are those where I'm utilizing an auto resize feature. I might move them later. The main reason I'm doing this is because I plan to have multiple resource files for various device configurations. This is also part of an attempt to fix an issue that's been reported, and I've previously been avoiding, where items are cut off on screens that are too small (#4).