Disabled missing translation check
Because missing translations actually cause the lint checks to fail, which in turn causes the builds to fail, I've disabled the check for them. It's not really a necessary check, and already exists in the IDE. Plus the absence of a full translation doesn't break anything. It's also worth noting that some translators might not actually find a need to translate certain strings. For instance, I doubt most people will bother translating the set titles that just contain letters. It's why I initially disabled those lines for translation, but in the case of spanish and catalan (and probably all romance languages), they need to be translatable because otherwise the words are in the wrong order. And that's why they were enabled by @laz08 (914cabb5ac47140000a2c014440777fc0ccf1ab5)(#6). So in conclusion, I need to disable the missing translation check for lint, while keeping it enabled for Android Studio, so I'm still aware of those situations.