Stopped LogView reload on rotate
Turns out both LogView and LogDetailView were reloading the entire activity whenever the screen was rotated, which was unnecessary because I've been using the onConfigurationChanged method to handle layout changes on screen rotation. The way to disable this is by adding an item to the manifest. So I did that.