Added hook for locale change
Set the vocabulary questions to reload whenever the locale changes. This is because, ideally, the new locale will have a different set of questions, assuming it's been translated. This is mainly to ensure the entire app is always in the same language, even if the user changes the language in system settings without reloading the app. Because before, if the user did change the language, all the UI text would be in the new language, but the vocabulary questions would remain in the previous language. This fixes that.