Moved question bank
Moved the storage of the quiz question bank into the QuestionManagement class, so it's not rebuilt every time the MainQuiz activity reloads. I also added a special hook to prevent the QuestionBank object from being rebuilt every time one so much as visits the Options or QuestionSelection screens. This hook stores a simple boolean array that acts as a record of all relevant user preferences associated with the current question bank. And if there is any change in that array, only then will it completely rebuild the question bank. This is mainly part of an attempt to fix the issue of the current question changing on rotate (#5). But more than that, I don't like things reloading unnecessarily. This system is more complicated, but there is certainly less overhead than completely rebuilding the question bank every time a change might've been possible, maybe.