Restricted the newQuestion calls
Redesigned how MainQuiz calls for new questions. I set it so it no longer does so every time the screen is reloaded. Now it will only generate a new question from the question bank if a question has just been answered, or the question bank has just been rebuilt (the relevant method now returns a boolean to indicate this). But as part of this, most of the code from the nextQuestion method had to be split off into a new refreshDisplay method. Because that code still needs to be run after the screen is reloaded, regardless of whether a new question has been called. I also had to change how the 'no questions' case is handled, because the code to deal with it has been moved to a separate method. So the newQuestion method in the QuestionBank class no longer throws a NoQuestionsException if there are no questions. It just does nothing, and allows other checks to deal with it. The result of all this means that the questions no longer change every time the screen is rotated, which should mean issue #5 has been fixed... except, the MultipleChoicePad still refreshes, which shouldn't happen. So I'll consider it partially fixed for now.