Continued changes from merge
I recently updated the kanji-questions branch with all the changes from the develop branch, in an attempt to ease merging when I finally finish building this feature. This is mainly because I needed to deal with file conflicts, and some of the changes I made in develop need to be continued in this branch. This includes the redesign in how digraph sections are marked in the xml files and how they're recorded (315d33a50039c53ef2257aa4b7bb64a05229321d), the transfer of hard-coded text sizes into the dimension resource file (70aa010f649114964c3534d29bf465e0e2027ea1), and the alteration in how the main question bank is stored and updated (4b33e0c472670f409da12391e595d61619f998a4). All of these changes needed to be applied to to the kanji question code, or updated to take the kanji questions into account.